python_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.10
h11==0.14.0
httpx==0.27.0
idna==3.10
pillow==11.1.0
pydantic==2.10.6
pydantic-core==2.27.2
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.46.0
typing-extensions==4.12.2
uvicorn==0.34.0
